+++ sdk/src/main/java/ly/count/android/sdk/ExplicitStorageCallback.java
@@ -1,0 +1,5 @@
+package ly.count.android.sdk;
+
+interface ExplicitStorageCallback {
+    void WriteToStorageFinished(boolean writeWasPerformed);
+}